requirements.txt
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
